pyproject.toml
[project]
name = "bdd_analysis"
version = "0.1.0"
requires-python = ">=3.12.0"
dependencies = [
  "black>=25.1.0",
  "dotenv>=0.9.9",
  "ipykernel>=6.29.5",
  "mlflow>=2.22.0",
  "requests>=2.32.3",
  "toml>=0.10.2",
  "torch>=2.7.0",
  "torchmetrics[detection]>=1.7.1",
  "torchvision>=0.22.0",
  "tqdm>=4.67.1",
  "ultralytics>=8.3.127",
  "opencv-python>=4.8.0.76",
  "pyyaml>=6.0",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
